feat(skillset): thread ProjectionScope through the projector contract
Task 2.2. Every entry point takes scope, defaulting to
ProjectionScope.everything(), so callers that declare nothing keep exactly
today's whole-set behaviour. Nothing reads it yet.

Moved ProjectionScope from _mutation_flow.py to
runtime_projection_contract.py. Two reasons: the flow module states it is
internal to the command services and not a public layer, while the scope is
part of the projection contract both protocols must declare; and
_mutation_flow already imports the contract, so the contract could not have
imported back without a cycle.

Also replaced a vendor product name in a comment I added last commit with
the neutral term -- core/ forbids it and
test_no_identity_vendor_in_core caught it. That is twice now an
architecture test has caught something my per-task runs missed, so those
now run with every task in this group, not just at group boundaries.
